+++ archiver/camonitorApp/tool_lib.h
@@ -79,7 +79,7 @@ typedef struct
     char firstStampPrinted;
     char onceConnected;
     evid evid;
-    long long callbackCounts;         //每次eventCallback调用时，将这个值加一。这个值可以用来统计不同PV的带宽占用比例。
+    int callbackCounts;         //每次eventCallback调用时，将这个值加一。这个值可以用来统计不同PV的带宽占用比例。
     int isConnected;
 } pv;
